extract error parsing logic into own module
