+++ app/src/lib/ssh/index.ts
@@ -1,1 +1,2 @@
 export { getSSHEnvironment } from './ssh-environment'
+export { isHostVerificationError } from './verification'
